pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catrax-pathfinder"
version = "1.0.2"
dependencies = [
    "numpy==1.26.4",
    "xgboost==2.1.4",
    "pandas==2.3.1",
    "mysql-connector-python",
    "seaborn",
    "matplotlib",
    "optuna",
]

description = "Pathfinder is designed to identify semantic paths between two biological entities."
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "Mohsen Taheri", email = "mohsen.taheri.1989@gmail.com" }
]
license = { text = "MIT" }        # or use "file" with LICENSE
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/Translator-CATRAX/pathfinder"
Issues = "https://github.com/Translator-CATRAX/pathfinder/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
# Include resource files in the wheel
"pathfinder.resources" = [
  "*.pkl",
  "pathfinder_xgboost_model_kg_2_10_2"
]
